requirements.txt
torch~=1.11.0+cu113
torchvision~=0.12.0+cu113
numpy~=1.21.5
einops~=0.6.0
timm~=0.6.11
pillow~=9.1.0
opencv-python~=4.6.0.66
tqdm~=4.64.1